+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * [#1897](https://github.com/bbatsov/rubocop/issues/1897): Don't report that semicolon separated statements can be converted to modifier form in `IfUnlessModifier` (and don't auto-correct them). ([@jonas054][])
 * [#1644](https://github.com/bbatsov/rubocop/issues/1644): Don't search the entire file system when a folder is named `,` (fix for jruby and rbx). ([@rrosenblum][])
 * [#1803](https://github.com/bbatsov/rubocop/issues/1803): Don't warn for `return` from `lambda` block in `NonLocalExitFromIterator`. ([@ypresto][])
+* [#1905](https://github.com/bbatsov/rubocop/issues/1905): Ignore sparse and trailing comments in Style/Documentation. ([@RGBD][])
 
 ## 0.31.0 (05/05/2015)
 
@@ -1427,3 +1428,4 @@
 [@chastell]: https://github.com/chastell
 [@glasnt]: https://github.com/glasnt
 [@crazydog115]: https://github.com/crazydog115
+[@RGBD]: https://github.com/RGBD